% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-crul.R
\name{build_crul_response}
\alias{build_crul_response}
\title{Build a crul response}
\usage{
build_crul_response(req, resp)
}
\arguments{
\item{req}{a request}

\item{resp}{a response}
}
\value{
a crul response
}
\description{
Build a crul response
}
\keyword{internal}
