% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-curl.R
\name{build_curl_request}
\alias{build_curl_request}
\title{Build a curl request}
\usage{
build_curl_request(x)
}
\arguments{
\item{x}{an unexecuted curl request object}
}
\value{
a curl request
}
\description{
Build a curl request
}
