% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-curl.R
\name{build_curl_response}
\alias{build_curl_response}
\title{Build a curl response}
\usage{
build_curl_response(req, resp)
}
\arguments{
\item{req}{a request}

\item{resp}{a response}
}
\value{
a curl response
}
\description{
Build a curl response
}
