% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-httr2.R
\name{build_httr2_request}
\alias{build_httr2_request}
\title{Build a httr2 request}
\usage{
build_httr2_request(x)
}
\arguments{
\item{x}{an unexecuted httr request object}
}
\value{
a httr request
}
\description{
Build a httr2 request
}
