% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-httr2.R
\name{build_httr2_response}
\alias{build_httr2_response}
\title{Build a httr2 response (\code{httr2_response})}
\usage{
build_httr2_response(req, resp)
}
\arguments{
\item{req}{a request}

\item{resp}{a response}
}
\value{
an httr2 response (\code{httr2_response})
}
\description{
Build a httr2 response (\code{httr2_response})
}
\keyword{internal}
