% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-httr.R
\name{build_httr_request}
\alias{build_httr_request}
\title{Build a crul request}
\usage{
build_httr_request(x)
}
\arguments{
\item{x}{an unexecuted crul request object}
}
\value{
a crul request
}
\description{
Build a crul request
}
