% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-httr.R
\name{build_httr_response}
\alias{build_httr_response}
\title{Build a httr response}
\usage{
build_httr_response(req, resp)
}
\arguments{
\item{req}{a request}

\item{resp}{a response}
}
\value{
a httr response
}
\description{
Build a httr response
}
\keyword{internal}
