% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-crul.R
\name{CrulAdapter}
\alias{CrulAdapter}
\title{CrulAdapter}
\description{
\pkg{crul} library adapter
}
\details{
This adapter modifies \pkg{crul} to allow mocking HTTP requests
}
\seealso{
Other http_lib_adapters: 
\code{\link{HttrAdapter}}
}
\concept{http_lib_adapters}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{adapter name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-enable}{\code{CrulAdapter$enable()}}
\item \href{#method-disable}{\code{CrulAdapter$disable()}}
\item \href{#method-handle_request}{\code{CrulAdapter$handle_request()}}
\item \href{#method-remove_crul_stubs}{\code{CrulAdapter$remove_crul_stubs()}}
\item \href{#method-clone}{\code{CrulAdapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enable"></a>}}
\subsection{Method \code{enable()}}{
Enable the adapter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrulAdapter$enable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE}, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disable"></a>}}
\subsection{Method \code{disable()}}{
Disable the adapter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrulAdapter$disable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{FALSE}, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-handle_request"></a>}}
\subsection{Method \code{handle_request()}}{
All logic for handling a request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrulAdapter$handle_request(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{a crul request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
various outcomes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_crul_stubs"></a>}}
\subsection{Method \code{remove_crul_stubs()}}{
Remove all crul stubs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrulAdapter$remove_crul_stubs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing returned; removes all crul request stubs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrulAdapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
