% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-curl.R
\name{curl_mock}
\alias{curl_mock}
\title{Turn on curl mocking}
\usage{
curl_mock(on = TRUE)
}
\arguments{
\item{on}{(logical) set to \code{TRUE} to turn on, and \code{FALSE}
to turn off. default: \code{TRUE}}
}
\value{
sets a env var to TRUE
}
\description{
Turn on curl mocking
}
