% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl_mocking.R
\name{curl_mock_req}
\alias{curl_mock_req}
\title{mock a curl request}
\usage{
curl_mock_req(url, h, path = NULL, called = "")
}
\arguments{
\item{url}{(character) a url, see \link[curl:curl_fetch]{curl::curl_fetch}}

\item{h}{a curl handle, see \code{\link[curl:new_handle]{curl::new_handle()}}}

\item{path}{a path. default: \code{NULL}. see \link[curl:curl_fetch]{curl::curl_fetch}}

\item{called}{the expression called, used in webmockr
internals}
}
\description{
mock a curl request
}
\examples{
library(curl)
h <- new_handle()
handle_setheaders(h, a = 'b')
handle_setheaders(h, foo = 'bar')
# curl_mock_req("https://httpbin.org/get", h, "")
stub_request("get", "https://httpbin.org/get")
curl_mock_req("https://httpbin.org/get", h, "")
}
