% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipswitch.R
\name{enable}
\alias{enable}
\alias{disable}
\title{Enable or disable webmockr}
\usage{
enable(options = list())

disable(options = list())
}
\arguments{
\item{options}{list of options}
}
\description{
Enable or disable webmockr
}
