% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handling.R
\name{handle_stub_removal}
\alias{handle_stub_removal}
\title{Handle stub removal}
\usage{
handle_stub_removal(.data, code)
}
\arguments{
\item{.data}{an object of class \code{StubbedRequest} required}

\item{code}{a code block. required}
}
\value{
if no error, the result of running \code{code}; if an error occurs
\code{\link[=withCallingHandlers]{withCallingHandlers()}} throws a warning and then the stub is removed
}
\description{
Handle stub removal
}
\keyword{internal}
