% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestPattern.R
\docType{data}
\name{HeadersPattern}
\alias{HeadersPattern}
\title{HeadersPattern}
\arguments{
\item{pattern}{(list) a pattern, as a named list, must be named}
}
\description{
HeadersPattern
}
\details{
\strong{Methods}
\describe{
\item{\code{matches(headers)}}{
Match a list of headers against that stored
- headers (list)
}
}
}
\examples{
x <- HeadersPattern$new(pattern = list(a = 5))
x$pattern
x$matches(list(a = 5))
}
\keyword{internal}
