% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-httr2.R
\name{httr2_mock}
\alias{httr2_mock}
\title{Turn on httr2 mocking}
\usage{
httr2_mock(on = TRUE)
}
\arguments{
\item{on}{(logical) \code{TRUE} to turn on, \code{FALSE} to turn off. default: \code{TRUE}}
}
\value{
Silently returns \code{TRUE} when enabled and \code{FALSE} when disabled.
}
\description{
Turn on httr2 mocking
}
