% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-httr.R
\docType{data}
\name{HttrAdapter}
\alias{HttrAdapter}
\title{httr library adapter}
\description{
httr library adapter
}
\details{
\strong{Methods}
\describe{
\item{\code{enable()}}{
Enable the adapter
}
\item{\code{disable()}}{
Disable the adapter
}
\item{\code{build_httr_request(x)}}{
Build a crul \link{RequestSignature}
x: crul request parts (list)
}
\item{\code{build_httr_response(req, resp)}}{
Build a crul response
req: a crul request (list)
resp: a crul response ()
}
\item{\code{handle_request()}}{
All logic for handling a request
req: a crul request (list)
}
\item{\code{remove_httr_stubs()}}{
Remove all crul stubs
}
}

This adapter modifies \pkg{httr} to allow mocking HTTP requests
}
\examples{
\dontrun{
# httr_mock(TRUE)
library(httr)
replay <- function(req) req
old_req <- set_callback("request", replay)
req <- GET("https://httpbin.org")
x <- HttrAdapter$new()
x$handle_request(req)

stub_request('get', uri = 'https://httpbin.org') \%>\%
  wi_th(
    headers = list('Accept' = 'application/json, text/xml, application/xml, */*')
)
stub_registry()
x$handle_request(req)

# another eg
library(httr)
replay <- function(req) {
  webmockr::HttrAdapter$new()$handle_request(req)
}
old_req <- set_callback("request", replay)
GET("https://httpbin.org/get")
stub_request('get', uri = 'https://httpbin.org/get') \%>\%
  wi_th(
    headers = list('Accept' = 'application/json, text/xml, application/xml, */*')
  ) \%>\%
  to_return(status = 200, body = "stuff", headers = list(a = 5))
GET("https://httpbin.org/get")
}
}
\seealso{
Other http_lib_adapters: \code{\link{CrulAdapter}}
}
\keyword{datasets}
