% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last_request}
\alias{last_request}
\title{Get the last HTTP request made}
\usage{
last_request()
}
\value{
\code{NULL} if no requests registered; otherwise the last
registered request made as a \code{RequestSignature} class
}
\description{
Get the last HTTP request made
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# no requests
request_registry_clear()
last_request()

# a request is found
enable()
stub_request("head", "https://nytimes.com")
library(crul)
crul::ok("https://nytimes.com")
last_request()

# cleanup
request_registry_clear()
stub_registry_clear()
\dontshow{\}) # examplesIf}
}
