% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_file.R
\name{mock_file}
\alias{mock_file}
\title{Mock file}
\usage{
mock_file(path, payload)
}
\arguments{
\item{path}{(character) a file path. required}

\item{payload}{(character) string to be written to the file given
at \code{path} parameter. required}
}
\value{
a list with S3 class \code{mock_file}
}
\description{
Mock file
}
\examples{
mock_file(path = tempfile(), payload = "{\"foo\": \"bar\"}")
}
