% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_body.R
\name{pluck_body}
\alias{pluck_body}
\title{Extract the body from an HTTP request}
\usage{
pluck_body(x)
}
\arguments{
\item{x}{an unexecuted crul \emph{or} httr request object}
}
\value{
one of the following:
\itemize{
\item \code{NULL} if the request is not associated with a body
\item \code{NULL} if an upload is used not in a list
\item list containing the multipart-encoded body
\item character vector with the JSON- or raw-encoded body, or upload form file
}
}
\description{
Returns an appropriate representation of the data contained within a request
body based on its encoding.
}
