% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_request_stub.R
\name{remove_request_stub}
\alias{remove_request_stub}
\title{Remove a request stub}
\usage{
remove_request_stub(stub)
}
\arguments{
\item{stub}{a request stub, of class \code{StubbedRequest}}
}
\value{
logical, \code{TRUE} if removed, \code{FALSE} if not removed
}
\description{
Remove a request stub
}
\examples{
(x <- stub_request("get", "https://httpbin.org/get"))
stub_registry()
remove_request_stub(x)
stub_registry()
}
\seealso{
Other stub-registry: 
\code{\link{StubRegistry}},
\code{\link{stub_registry_clear}()},
\code{\link{stub_registry}()}
}
\concept{stub-registry}
