% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestSignature.R
\docType{data}
\name{RequestSignature}
\alias{RequestSignature}
\title{General purpose request signature builder}
\arguments{
\item{method}{the HTTP method (any, head, options, get, post, put,
patch, trace, or delete). "any" matches any HTTP method. required.}

\item{uri}{(character) request URI. required.}

\item{options}{(list) options. optional}
}
\description{
General purpose request signature builder
}
\details{
\strong{Methods}
\describe{
\item{\code{to_s()}}{
Request signature to a string
}
}
}
\examples{
# make request signature
x <- RequestSignature$new(method = "get", uri = "https:/httpbin.org/get")
# method
x$method
# uri
x$uri
# request signature to string
x$to_s()
}
\keyword{datasets}
