% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\docType{data}
\name{Response}
\alias{Response}
\title{Response class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Response
}
\arguments{
\item{options}{(list) a list of options}
}
\description{
Response class
}
\details{
\strong{Methods}
  \describe{
    \item{\code{set_request_headers(headers)}}{
      set request headers
      - headers: a list of key-value pair headers
    }
    \item{\code{get_request_headers()}}{
      get request headers
    }
    \item{\code{set_response_headers(headers)}}{
      set response headers
      - headers: a list of key-value pair headers
    }
    \item{\code{get_response_headers()}}{
      get response headers
    }
    \item{\code{set_body(body)}}{
      - body: must be a string
    }
    \item{\code{get_body()}}{
      get body
    }
    \item{\code{set_status()}}{
      - body: must be an integer status code
    }
    \item{\code{get_status()}}{
      get status code
    }
    \item{\code{set_exception()}}{
      set exception
    }
    \item{\code{get_exception()}}{
      get exception
    }
  }
}
\examples{
\dontrun{
(x <- Response$new())

x$set_url("https://httpbin.org/get")
x

x$set_request_headers(list('Content-Type' = "application/json"))
x
x$request_headers

x$set_response_headers(list('Host' = "httpbin.org"))
x
x$response_headers

x$set_status(404)
x
x$get_status()

x$set_body("hello world")
x
x$get_body()

x$set_exception("exception")
x
x$get_exception()
}
}
\keyword{internal}
