% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{save_local}
\alias{save_local}
\title{Save results locally}
\usage{
save_local(x, y, path = "~/")
}
\arguments{
\item{x}{Output from API call}

\item{y}{Cache key}
}
\description{
Save results locally
}
\keyword{internal}

