% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub_registry.R
\name{stub_registry}
\alias{stub_registry}
\title{List stubs in the stub registry}
\usage{
stub_registry()
}
\value{
an object of class \code{StubRegistry}, print method gives the
stubs in the registry
}
\description{
List stubs in the stub registry
}
\seealso{
Other stub-registry: \code{\link{StubRegistry}},
  \code{\link{remove_request_stub}},
  \code{\link{stub_registry_clear}}
}
