% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StubbedRequest.R
\docType{data}
\name{StubbedRequest}
\alias{StubbedRequest}
\title{StubbedRequest class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StubbedRequest
}
\arguments{
\item{method}{the HTTP method (any, head, options, get, post, put,
patch, trace, or delete). "any" matches any HTTP method. required.}

\item{uri}{(character) request URI. required.}

\item{body}{(list) request body, as a list. optional}

\item{query}{(list) query parameters, as a list. optional}

\item{headers}{(list) request headers. optional.}
}
\description{
StubbedRequest class
}
\details{
\strong{Methods}
  \describe{
    \item{\code{with()}}{
      with slots for:
      \itemize{
       \item status
       \item body
       \item headers
      }
    }
    \item{\code{to_return()}}{
      Stubbed response
    }
  }
}
\examples{
\dontrun{
x <- StubbedRequest$new(method = "get", uri = "api.crossref.org")
x$method
x$uri
x$with(headers = list('User-Agent' = 'R'))
x$to_return(status = 200, body = "foobar", headers = list())
x$to_s()
}
}
\keyword{internal}

