% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_timeout.R
\name{to_timeout}
\alias{to_timeout}
\title{Set timeout as an expected return on a match}
\usage{
to_timeout(.data)
}
\arguments{
\item{.data}{input. Anything that can be coerced to a \code{StubbedRequest} class
object}
}
\value{
an object of class \code{StubbedRequest}, with print method describing
the stub
}
\description{
Set timeout as an expected return on a match
}
\note{
see examples in \code{\link[=stub_request]{stub_request()}}
}
