% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestPattern.R
\docType{data}
\name{UriPattern}
\alias{UriPattern}
\title{UriPattern}
\arguments{
\item{pattern}{(character) a uri}
}
\description{
UriPattern
}
\details{
\strong{Methods}
\describe{
\item{\code{matches(uri)}}{
Match a uri against that given in \code{pattern}
- uri (character) a uri
}
}
}
\examples{
z <- UriPattern$new(pattern = "http://foobar.com")
z$matches("http://foobar.com")
}
\keyword{internal}
