% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmockr.R
\docType{package}
\name{webmockr}
\alias{webmockr}
\alias{webmockr-package}
\title{Stubbing and setting expectations on HTTP requests}
\description{
Stubbing and setting expectations on HTTP requests
}
\keyword{package}

