% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{webmockr_crul_fetch}
\alias{webmockr_crul_fetch}
\title{execute a curl request}
\usage{
webmockr_crul_fetch(x)
}
\arguments{
\item{x}{an object}
}
\value{
a curl response
}
\description{
execute a curl request
}
\keyword{internal}
