% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmockr_reset.R
\name{webmockr_reset}
\alias{webmockr_reset}
\title{webmockr_reset}
\usage{
webmockr_reset()
}
\value{
nothing
}
\description{
Clear all stubs and the request counter
}
\details{
this function runs \code{\link[=stub_registry_clear]{stub_registry_clear()}} and
\code{\link[=request_registry_clear]{request_registry_clear()}} - so you can run those two yourself
to achieve the same thing
}
\examples{
# webmockr_reset()
}
\seealso{
\code{\link[=stub_registry_clear]{stub_registry_clear()}} \code{\link[=request_registry_clear]{request_registry_clear()}}
}
