% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestPattern.R
\docType{data}
\name{BodyPattern}
\alias{BodyPattern}
\title{BodyPattern}
\arguments{
\item{pattern}{(list) a body object}
}
\description{
BodyPattern
}
\details{
\strong{Methods}
\describe{
\item{\code{matches(body, content_type = "")}}{
Match a body object against that given in \code{pattern}
- body (list) the body
- content_type (character) content type
}
}
}
\examples{
# make a request signature
bb <- RequestSignature$new(
  method = "get",
  uri = "https:/httpbin.org/get",
  options = list(
    body = list(foo = "bar", a = 5)
  )
)

# make body pattern object
z <- BodyPattern$new(pattern = list(foo = "bar"))
z$pattern
z$matches(bb$body)
}
\keyword{internal}
