% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-httr2.R
\name{build_httr2_request}
\alias{build_httr2_request}
\title{Build an httr2 request}
\usage{
build_httr2_request(x)
}
\arguments{
\item{x}{an unexecuted httr2 request object}
}
\value{
a \code{httr2_request}
}
\description{
Build an httr2 request
}
