% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-crul.R
\docType{data}
\name{CrulAdapter}
\alias{CrulAdapter}
\title{crul library adapter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CrulAdapter
}
\arguments{
\item{x}{input}
}
\description{
crul library adapter
}
\details{
This adapter modifies \pkg{crul} to allow mocking HTTP requests
when one is using \pkg{crul} in their code
}
\seealso{
Other http_lib_adapters: \code{\link{HttrAdapter}}
}
\keyword{datasets}
