% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipswitch.R
\name{enable}
\alias{enable}
\alias{enabled}
\alias{disable}
\title{Enable or disable webmockr}
\usage{
enable(adapter = NULL, options = list(), quiet = FALSE)

enabled(adapter = "crul")

disable(adapter = NULL, options = list(), quiet = FALSE)
}
\arguments{
\item{adapter}{(character) the adapter name, 'crul' or 'httr'.
one or the other. if none given, we attempt to enable both
adapters}

\item{options}{list of options - ignored for now.}

\item{quiet}{(logical) suppress messages? default: \code{FALSE}}
}
\value{
\code{enable()} and \code{disable()} invisibly returns booleans for
each adapter, as a result of running enable or disable, respectively,
on each \link{HttpLibAdapaterRegistry} object. \code{enabled} returns a
single boolean
}
\description{
Enable or disable webmockr
}
\details{
\code{enable()} enables \pkg{webmockr} for all adapters.
\code{disable()} disables \pkg{webmockr} for all adapters.  \code{enabled()}
answers whether \pkg{webmockr} is enabled for a given adapter
}
