% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestRegistry.R
\docType{data}
\name{HashCounter}
\alias{HashCounter}
\title{hash with counter, to store requests, and count each time it is used}
\description{
hash with counter, to store requests, and count each time it is used
}
\details{
\strong{Methods}
\describe{
\item{\code{put(key)}}{
Register a request by it's key
- key: a character string of the request, serialized from
\link{CrulAdapter} or other adapter
}
\item{\code{get(key)}}{
Get a request by key
- key: a character string of the request, serialized from
\link{CrulAdapter} or other adapter
}
}
}
\examples{
x <- HashCounter$new()
x$put("foo bar")
x$put("foo bar")
x$put("hello world")
x$put("hello world")
x$put("hello world")
x$hash
}
\seealso{
Other request-registry: \code{\link{RequestRegistry}},
  \code{\link{request_registry}}
}
\keyword{datasets}
