% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HttpLibAdapterRegistry.R
\docType{data}
\name{HttpLibAdapaterRegistry}
\alias{HttpLibAdapaterRegistry}
\title{http lib adapter registry}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HttpLibAdapaterRegistry
}
\description{
http lib adapter registry
}
\examples{
x <- HttpLibAdapaterRegistry$new()
x$register(HttrAdapter$new())
x
x$adapters
x$adapters[[1]]$name
}
\keyword{datasets}
