% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HttpLibAdapterRegistry.R
\docType{data}
\name{HttpLibAdapaterRegistry}
\alias{HttpLibAdapaterRegistry}
\title{http lib adapter registry}
\description{
http lib adapter registry
}
\details{
\strong{Methods}
\describe{
\item{\code{register(x)}}{
Register an http library adapter
x: an http lib adapter, e.g., \link{CrulAdapter}
return: nothing, registers the library adapter
}
}
}
\examples{
x <- HttpLibAdapaterRegistry$new()
x$register(CrulAdapter$new())
x
x$adapters
x$adapters[[1]]$name
}
\keyword{datasets}
