% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-httr.R
\name{httr_mock}
\alias{httr_mock}
\title{Turn on \code{httr} mocking}
\usage{
httr_mock(on = TRUE)
}
\arguments{
\item{on}{(logical) set to \code{TRUE} to turn on, and \code{FALSE}
to turn off. default: \code{TRUE}}
}
\value{
Silently returns \code{TRUE} when enabled and \code{FALSE} when disabled.
}
\description{
Sets a callback that routes \code{httr} requests through \code{webmockr}
}
