% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr-adapter.R
\docType{data}
\name{HttrAdapter}
\alias{HttrAdapter}
\title{httr library adapter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HttrAdapter
}
\arguments{
\item{x}{input}
}
\description{
httr library adapter
}
\details{
This adapter modifies \pkg{httr} to allow mocking HTTP requests
when one is using \pkg{httr} in their code
}
\keyword{internal}

