% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/including.R
\name{including}
\alias{including}
\title{partially match}
\usage{
including(x)
}
\arguments{
\item{x}{input, a list}
}
\description{
partially match
}
\examples{
including(list("foo"))
including(list(foo = "bar"))
# just keys by setting values as NULL
including(list(foo = NULL, bar = NULL))

# in a stub
req <- stub_request("get", "https://httpbin.org/get")
wi_th(req, body = list(foo = "bar"))
wi_th(req, body = including(list(foo = "bar")))
}
