% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_exclude.R
\name{including}
\alias{including}
\alias{excluding}
\title{Partially match query parameters or request bodies}
\usage{
including(x)

excluding(x)
}
\arguments{
\item{x}{various objects}
}
\value{
same as \code{x}, but with attributes added
}
\description{
Partially match query parameters or request bodies
}
\examples{
including(list(foo = "bar"))
# just keys by setting values as NULL
including(list(foo = NULL, bar = NULL))

# in a stub
req <- stub_request("get", "https://httpbin.org/get")

## query
wi_th(req, query = list(foo = "bar"))
wi_th(req, query = including(list(foo = "bar")))
wi_th(req, query = excluding(list(foo = "bar")))

# cleanup
stub_registry_clear()
}
