% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestPattern.R
\docType{data}
\name{MethodPattern}
\alias{MethodPattern}
\title{MethodPattern}
\arguments{
\item{pattern}{(character) a HTTP method, lowercase}
}
\description{
MethodPattern
}
\details{
\strong{Methods}
\describe{
\item{\code{matches(method)}}{
An HTTP method
- method (character)
}
}
}
\examples{
x <- MethodPattern$new(pattern = "post")
x$pattern
x$matches(method = "post")
}
\keyword{internal}
