% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestRegistry.R
\name{request_registry_filter}
\alias{request_registry_filter}
\title{Request registry filter}
\usage{
request_registry_filter(method = NULL, url = NULL, body = NULL)
}
\arguments{
\item{method}{(character) http method}

\item{url}{(character) a url}

\item{body}{(character) a request body}
}
\value{
list, length of number of unique requests recorded, or
those requests matching filters supplied by parameters
}
\description{
If not filters are given, returns all requests
}
\examples{
enable()

stub_request("any", uri_regex = ".+")

library(httr)
GET("http://example.com")
GET("https://hb.cran.dev/get")
POST("https://hb.cran.dev/post", body = list(fruit = "apple"))
POST("https://hb.cran.dev/post", body = list(cheese = "swiss"))

request_registry_filter()
request_registry_filter(method="get")
request_registry_filter(method="post")
request_registry_filter(method="get", url="http://example.com")
request_registry_filter(method="post", body=list(fruit = "apple"))
request_registry_filter(method="post", body=list(cheese = "swiss"))
request_registry_filter(method="post", body=list(cheese = "cheddar"))

match <- request_registry_filter(method="post")[[1]]
match$request
match$request$to_s()
match$count
}
