% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestPattern.R
\docType{data}
\name{RequestPattern}
\alias{RequestPattern}
\title{RequestPattern class}
\arguments{
\item{method}{the HTTP method (any, head, options, get, post, put,
patch, trace, or delete). "any" matches any HTTP method. required.}

\item{uri}{(character) request URI. required.}

\item{options}{(list) options. optional}
}
\description{
RequestPattern class
}
\details{
\strong{Methods}
\describe{
\item{\code{matches(request_signature)}}{
Test if request_signature matches a pattern
- request_signature: a request signature
}
\item{\code{to_s()}}{
Print pattern for easy human consumption
}
}
}
\examples{
\dontrun{
(x <- RequestPattern$new(method = "get", uri = "https://httpbin.org/get"))
x$body_pattern
x$headers_pattern
x$method_pattern
x$uri_pattern
x$to_s()

# make a request signature
rs <- RequestSignature$new(method = "get", uri = "https://httpbin.org/get")

# check if it matches
x$matches(rs)
}
}
\keyword{datasets}
