% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestRegistry.R
\docType{data}
\name{RequestRegistry}
\alias{RequestRegistry}
\title{Request registry}
\description{
Request registry
}
\details{
\strong{Methods}
\describe{
\item{\code{register_request(request)}}{
Register a request
- request: a character string of the request, serialized from
\link{CrulAdapter} or other adapter
}
\item{\code{reset()}}{
Reset the registry to no registered requests
}
}
}
\examples{
x <- RequestRegistry$new()
x$register_request(request = "GET http://scottchamberlain.info")
x$register_request(request = "GET http://scottchamberlain.info")
x$register_request(request = "POST https://httpbin.org/post")
# print method to list requests
x

# hashes, and number of times each requested
x$request_signatures$hash

# reset the request registry
x$reset()
}
\seealso{
Other request-registry: \code{\link{HashCounter}},
  \code{\link{request_registry}}
}
\concept{request-registry}
\keyword{datasets}
