% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestSignature.R
\docType{data}
\name{RequestSignature}
\alias{RequestSignature}
\title{general purpose request signature builder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RequestSignature
}
\description{
general purpose request signature builder
}
\keyword{internal}

