% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub_registry.R
\name{stub_registry}
\alias{stub_registry}
\title{List stubs in the stub registry}
\usage{
stub_registry()
}
\value{
an object of class \code{StubRegistry}, print method gives the
stubs in the registry
}
\description{
List stubs in the stub registry
}
\examples{
# make a stub
stub_request("get", "https://httpbin.org/get") \%>\%
  to_return(body = "success!", status = 200)

# check the stub registry, there should be one in there
stub_registry()

# make another stub
stub_request("get", "https://httpbin.org/get") \%>\%
  to_return(body = "woopsy", status = 404)

# check the stub registry, now there are two there
stub_registry()

# to clear the stub registry
stub_registry_clear()
}
\seealso{
Other stub-registry: 
\code{\link{StubRegistry}},
\code{\link{remove_request_stub}()},
\code{\link{stub_registry_clear}()}
}
\concept{stub-registry}
