% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub_registry_clear.R
\name{stub_registry_clear}
\alias{stub_registry_clear}
\title{Clear the stub registry}
\usage{
stub_registry_clear()
}
\value{
nothing, well technically an empty list, but
it's not anything useful
}
\description{
Clear all stubs
}
\examples{
(x <- stub_request("get", "https://httpbin.org/get"))
stub_registry()
stub_registry_clear()
stub_registry()
}
\seealso{
Other stub-registry: \code{\link{StubRegistry}},
  \code{\link{remove_request_stub}},
  \code{\link{stub_registry}}
}
