% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub_request.R
\name{stub_request}
\alias{stub_request}
\title{Stub an http request}
\usage{
stub_request(method = "get", url)
}
\arguments{
\item{method}{(character) HTTP method, one of "get", "post", "put", "patch",
"head", "delete", "options" - or the special "any" (for any method)}

\item{url}{(character) The request url. Can be a full url, partial, or a
regular expression to match many incantations of a url}
}
\description{
Stub an http request
}
\examples{
\dontrun{
# URI regex
stub_request("get", ".*google.*")

# add header
stub_request("get", ".*google.*")

# add expectation with to_return
stub_request("get", ".*google.*")
}
}

