% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub_request.R
\name{stub_request}
\alias{stub_request}
\title{Stub an http request}
\usage{
stub_request(method = "get", url)
}
\arguments{
\item{method}{(character) HTTP method, one of "get", "post", "put", "patch",
"head", "delete", "options" - or the special "any" (for any method)}

\item{url}{(character) The request url. Can be a full url, partial, or a
regular expression to match many incantations of a url}
}
\description{
Stub an http request
}
\examples{
\dontrun{
# URI regex
stub_request("get", ".*google.*")

# add header
stub_request("get", ".*google.*")

# add expectation with to_return
stub_request("get", ".*google.*")
}
\dontrun{
# basic stubbing
stub_request("get", url="http://httpbin.org/get")
stub_request("post", url="http://httpbin.org/post")

# add header
stub_request("get", url="http://httpbin.org/get") \%>\%
   wi_th(headers = list('User-Agent' = 'R'))

# add expectation with to_return
stub_request("get", url="http://httpbin.org/get") \%>\%
  wi_th(query = list(hello = "world"), headers = list('User-Agent' = 'R')) \%>\%
  to_return(status = 200, body = "", headers = list())

# RFC 6570 templates
stub_request("get", "www.example.com/{id}/")
stub_request("get", "/.*example.*/")
stub_request("get", "www.example.com/thing/{id}.json{?x,y,z}{&other*}")
# httr::GET('www.example.com/thing/5.json?x=1&y=2&z=3&anyParam=4')
}
}
