% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub_request.R
\name{stub_request}
\alias{stub_request}
\title{Stub an http request}
\usage{
stub_request(verb = "get", url)
}
\arguments{
\item{verb}{curl verb, one of get, post, put, patch, head, delete.}

\item{url}{Base url for the stub reqeust}

\item{where}{Where to store stubbed request. Only tests/cache/ for now.}
}
\description{
Stub an http request
}
\examples{
\donttest{
stub_request("get", url="http://google.com/")
GET("http://google.com/")
}
}

