% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StubbedRequest.R
\docType{data}
\name{StubbedRequest}
\alias{StubbedRequest}
\title{StubbedRequest class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StubbedRequest
}
\arguments{
\item{method}{the HTTP method (i.e. :head, :options, :get, :post, :put, :patch or :delete)}

\item{uri}{the request URI}

\item{body}{the request body}

\item{headers}{the request headers}
}
\description{
StubbedRequest class
}
\details{
\strong{Methods}
  \describe{
    \item{\code{to_hash()}}{
      Create a hash.
    }
    \item{\code{from_hash()}}{
      Get a hash back to an R list.
    }
  }
}
\examples{
\dontrun{
x <- StubbedRequest$new(method = "get", uri = "api.crossref.org")
x$method
x$uri
x$with(headers = list('User-Agent' = 'R'))
x$to_return(status = 200, body = "", headers = list())
}
}
\keyword{internal}

