% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StubRegistry.R
\docType{data}
\name{StubRegistry}
\alias{StubRegistry}
\alias{stub_registry}
\title{Stub registry}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StubRegistry

stub_registry()
}
\arguments{
\item{stub}{an object of class \code{StubbedRequest}}
}
\description{
Stub registry
}
\details{
\strong{Methods}
  \describe{
    \item{\code{register_stub(stub)}}{
      Register a stub
    }
    \item{\code{find_stubbed_request()}}{
      Find a stubbed request
    }
  }
}
\examples{
\dontrun{
# an actual stub
stub1 <- StubbedRequest$new(method = "get", uri = "api.crossref.org")
stub1$with(request_headers = list('User-Agent' = 'R'))
stub1$to_return(status = 200, body = "foobar", response_headers = list())
stub1

stub2 <- StubbedRequest$new(method = "get", uri = "api.crossref.org")
stub2

reg <- StubRegistry$new()
reg$register_stub(stub = stub1)
reg$register_stub(stub = stub2)
reg
reg$request_stubs
}
}
\keyword{internal}
