% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_return.R
\name{to_return}
\alias{to_return}
\alias{to_return_}
\title{Expection for a stubbed request}
\usage{
to_return(.data, ...)

to_return_(.data, ..., .dots)
}
\arguments{
\item{.data}{input. Anything that can be coerced to a \code{StubbedRequest} class
object}

\item{...}{Comma separated list of unquoted variable names, passed on
to \code{\link[lazyeval:lazy_dots]{lazyeval::lazy_dots()}}. accepts the following: status, body,
headers}

\item{.dots}{Used to work around non-standard evaluation}
}
\value{
an object of class \code{StubbedRequest}, with print method describing
the stub
}
\description{
Expection for a stubbed request
}
