% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_return.R
\name{to_return}
\alias{to_return}
\alias{to_return_}
\title{Expectation for what's returned from a stubbed request}
\usage{
to_return(.data, ...)

to_return_(.data, ..., .dots)
}
\arguments{
\item{.data}{input. Anything that can be coerced to a \code{StubbedRequest} class
object}

\item{...}{Comma separated list of variable names, passed on
to \code{\link[lazyeval:lazy_dots]{lazyeval::lazy_dots()}}. accepts the following: status, body,
headers. See Details for more.}

\item{.dots}{Used to work around non-standard evaluation}
}
\value{
an object of class \code{StubbedRequest}, with print method describing
the stub
}
\description{
Set response status code, response body, and/or response headers
}
\details{
Values for status, body, and headers:
\itemize{
\item status: (numeric/integer) three digit status code
\item body: various, including character string, list, raw, numeric, etc
\item headers: (list) a named list
}
}
\note{
see examples in \code{\link[=stub_request]{stub_request()}}
}
