% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{webmockr-defunct}
\alias{webmockr-defunct}
\title{Defunct functions in webmockr}
\description{
\itemize{
\item \code{\link[=webmockr_enable]{webmockr_enable()}}: Function removed, see \code{\link[=enable]{enable()}}
\item \code{\link[=webmockr_disable]{webmockr_disable()}}: Function removed, see \code{\link[=disable]{disable()}}
}
}
