% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmockr.R
\docType{package}
\name{webmockr-package}
\alias{webmockr-package}
\alias{webmockr}
\title{webmockr}
\description{
Stubbing and setting expectations on HTTP requests
}
\section{Features}{

\itemize{
\item Stubbing HTTP requests at low http client lib level
\item Setting and verifying expectations on HTTP requests
\item Matching requests based on method, URI, headers and body
\item Supports multiple HTTP libraries, including \pkg{crul},
\pkg{httr}, and \pkg{httr2}
\item Integration with HTTP test caching library \pkg{vcr}
}
}

\examples{
library(webmockr)
stub_request("get", "https://httpbin.org/get")
stub_request("post", "https://httpbin.org/post")
stub_registry()
}
\author{
Scott Chamberlain \email{myrmecocystus+r@gmail.com}

Aaron Wolen
}
\keyword{package}
