% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmockr-opts.R
\name{webmockr_configure}
\alias{webmockr_allow_net_connect}
\alias{webmockr_configuration}
\alias{webmockr_configure}
\alias{webmockr_configure_reset}
\alias{webmockr_disable}
\alias{webmockr_disable_net_connect}
\alias{webmockr_enable}
\alias{webmockr_net_connect_allowed}
\title{webmockr configuration}
\usage{
webmockr_configure(turn_on = FALSE, allow_net_connect = FALSE,
  allow_localhost = FALSE, allow = FALSE,
  net_http_connect_on_start = FALSE, show_stubbing_instructions = FALSE,
  query_values_notation = FALSE, show_body_diff = FALSE)

webmockr_configure_reset()

webmockr_configuration()

webmockr_enable()

webmockr_disable()

webmockr_allow_net_connect()

webmockr_disable_net_connect()

webmockr_net_connect_allowed()
}
\arguments{
\item{turn_on}{(logical) Default: \code{FALSE}}

\item{allow_net_connect}{(logical) Default: \code{TRUE}}

\item{allow_localhost}{(logical) Default: \code{TRUE}}

\item{allow}{(logical) Default: \code{TRUE}}

\item{net_http_connect_on_start}{(logical) Default: \code{TRUE}}

\item{show_stubbing_instructions}{(logical) Default: \code{TRUE}}

\item{query_values_notation}{(logical) Default: \code{TRUE}}

\item{show_body_diff}{(logical) Default: \code{TRUE}}
}
\description{
webmockr configuration
}
\examples{
\dontrun{
webmockr_configure()
webmockr_configure(
 allow_localhost = TRUE
)
webmockr_configuration()
webmockr_configure_reset()

webmockr_allow_net_connect()
webmockr_net_connect_allowed()
webmockr_disable_net_connect()
webmockr_net_connect_allowed()
}
}

