% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wi_th.R
\name{wi_th}
\alias{wi_th}
\alias{wi_th_}
\title{Set additional parts of a stubbed request}
\usage{
wi_th(.data, ...)

wi_th_(.data, ..., .dots)
}
\arguments{
\item{.data}{input. Anything that can be coerced to a \code{StubbedRequest} class
object}

\item{...}{Comma separated list of variable names, passed on
to \code{\link[lazyeval:lazy_dots]{lazyeval::lazy_dots()}}. accepts the following: query, body,
headers}

\item{.dots}{Used to work around non-standard evaluation}
}
\value{
an object of class \code{StubbedRequest}, with print method describing
the stub
}
\description{
Set query params, request body, and/or request headers
}
\details{
\code{with} is a function in the \code{base} package, so we went with
\code{wi_th}
}
\note{
see examples in \code{\link[=stub_request]{stub_request()}}
}
