% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wi_th.R
\name{wi_th}
\alias{wi_th}
\alias{wi_th_}
\title{Additional inputs to a stubbed request}
\usage{
wi_th(.data, ...)

wi_th_(.data, ..., .dots)
}
\arguments{
\item{.data}{input. Anything that can be coerced to a StubbedRequest class
object}

\item{...}{Comma separated list of unquoted variable names, passed on
to \code{\link[lazyeval]{lazy_dots}}}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
Additional inputs to a stubbed request
}

